requirements.txt
Flask
python-dotenv
spacy
serpapi
openai
google-search-results